Splitting TorchTensor.cs into multiple files, organized by functionality.
